# M2-Challenge-Tessema-Elleni/month-and-math/src/main/java/monthandmathservice/model/CustomErrorResponse.java
package monthandmathservice.model;

import com.fasterxml.jackson.annotation.JsonFormat;
import org.omg.CORBA.PUBLIC_MEMBER;

import java.time.LocalDateTime;

public class CustomErrorResponse {

    String  errorMsg;
    int status;
    String errorCode;
    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-mm-dd hh:mm:ss")
    LocalDateTime timestamp;

    public CustomErrorResponse() {

    }
    public CustomErrorResponse(String errorCode, String errorMsg) {
        this.errorCode = errorCode;
        this.errorMsg = errorMsg;
    }

    public String getErrorMsg() {
        return errorMsg;
    }

    public void setErrorMsg(String errorMsg) {
        this.errorMsg = errorMsg;
    }

    public int getStatus() {
        return status;
    }

    public void setStatus(int status) {
        this.status = status;
    }

    public String getErrorCode() {
        return errorCode;
    }

    public void setErrorCode(String errorCode) {
        this.errorCode = errorCode;
    }

    public LocalDateTime getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(LocalDateTime timestamp) {
        this.timestamp = timestamp;
    }
}